requirements.txt
psycopg2-binary==2.9.9
pydantic==2.5.0
boto3==1.28.65
botocore==1.31.65
pytest==7.4.3
pytest-cov==4.1.0
black==23.10.1
isort==5.12.0
mypy==1.6.1
flake8==6.1.0
python-dotenv==1.0.0
pytest=8.3.5
